アクセス制限を増加

+++ routes/verify/[token].tsx
@@ -13,7 +13,14 @@ export const handler: Handlers = {
     if (!check) {
       return ctx.render();
     }
-    await Verify.deleteMany({ token: token });
+    if (check.times > 9) {
+      await Verify.deleteMany({ token: token });
+    } else {
+      check.times++;
+      await Verify.updateMany({ token: token }, {
+        $set: { times: check.times },
+      });
+    }
     return ctx.render(check);
   },
 };
